intro level coming along
I am making it so that each individual level can be enhanced by running
it from a derrived version of LMap. In this case I am derriving
tutorialLevel from LMap, in which I add doors that open after a
condition is met. I am also working on making simple turrets to fight
against in the level which is coming along.